% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_retrieve.R
\name{wos_retrieve}
\alias{wos_retrieve}
\alias{wos_retrieve_all}
\title{Retrieve records from a search query result}
\usage{
wos_retrieve(result, first = 1, count = 100)

wos_retrieve_all(result)
}
\arguments{
\item{result}{a search query result generated by a previous call to \code{\link{wos_search}}}

\item{first}{index of the firest record to retrieve}

\item{count}{number of records to retrieve}
}
\value{
A tibble (data frame) with the records results. For fields with multiple value, such as authors or keywords, values are returne as a concatenated string.
}
\description{
Retrieve records from a search query result
}
\section{Functions}{
\itemize{
\item \code{wos_retrieve_all}: Retrieve all records from a search query result
}}

