% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_retrieve.R
\name{wos_retrieve_page}
\alias{wos_retrieve_page}
\title{Retrieve a batch of records (100 max)}
\usage{
wos_retrieve_page(sid, result, first = NULL, count = 100)
}
\description{
Retrieve a batch of records (100 max)
}
