% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_retrieve.R
\name{wos_retrieve}
\alias{wos_retrieve}
\alias{wos_retrieve_all}
\title{Retrieve records from a search query result}
\usage{
wos_retrieve(sid, result, first = NULL, number = 100)

wos_retrieve_all(sid, result)
}
\arguments{
\item{sid}{session id generated by a previous call to \code{\link{wos_authenticate}}}

\item{result}{a search query result generated by a previous call to \code{\link{wos_search}}}

\item{first}{index of the firest record to retrieve}

\item{number}{number of records to retrieve}
}
\value{
A tibble (data frame) with the records results. For fields with multiple value, such as authors or keywords, values are returne as a concatenated string.
}
\description{
Retrieve records from a search query result
}
\section{Functions}{
\itemize{
\item \code{wos_retrieve_all}: Retrieve all records from a search query result
}}

