% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_retrieve.R
\name{wos_retrieve_page}
\alias{wos_retrieve_page}
\title{Retrieve a batch of records (100 max)}
\usage{
wos_retrieve_page(result, first = NULL, count = 100)
}
\arguments{
\item{result}{a search query result generated by a previous call to \code{\link{wos_search}}}

\item{first}{index of the firest record to retrieve}

\item{count}{number of records to retrieve}
}
\description{
Retrieve a batch of records (100 max)
}
