% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wos_search.R
\name{wos_search}
\alias{wos_search}
\title{Run a search query to the Web of Science Web Services Lite API}
\usage{
wos_search(sid, query = "")
}
\arguments{
\item{sid}{session id generated by a previous call to \code{\link{wos_authenticate}}}

\item{query}{query text}
}
\value{
a list with two elements : the number of results, and the generated query id.
}
\description{
Run a search query to the Web of Science Web Services Lite API
}
\seealso{
\code{\link{wos_authenticate}}, \code{\link{wos_retrieve}}
}
